Update 01-Web-of-Data.md
Updated the Selected External Lesson Material under "Benefits of linked open data with examples". The original link to the University of Southampton's "Introduction to Linked Data and the Semantic Web" is not running and won’t be running anytime soon. I contacted one of the course instructors Dr Luis-Daniel Ibáñez and he suggested a few alternative online courses that are relevant and active.